+++ .eslintrc.js
@@ -39,7 +39,6 @@ module.exports = {
         "no-shadow": 1,
         "no-underscore-dangle": ["error", { "allow": ["_requestType", "_ptfmRequestType"] }],
         "no-use-before-define": 1,
-        "no-var": 0,
         "object-shorthand": [2, "consistent"],
         "one-var": 0,
         "prefer-destructuring": 0,
@@ -53,6 +52,5 @@ module.exports = {
             "requireParamDescription": false,
             "requireReturnDescription": false
         }],
-        "vars-on-top": 0
     }
 };